Migrated breadcrumbs and all bucket components to functional components. Added rjsf types.
