+++ src/components/Breadcrumbs/index.ts
@@ -1,2 +1,2 @@
-import { Breadcrumbs } from "./Breadcrumbs";
+import Breadcrumbs from "./Breadcrumbs";
 export default Breadcrumbs;
