+++ src/components/bucket/BucketCreate.tsx
@@ -1,6 +1,5 @@
+import React from "react";
 import type { SessionState, BucketState } from "../../types";
-
-import React, { PureComponent } from "react";
 
 import * as BucketActions from "../../actions/bucket";
 import BucketForm from "./BucketForm";
@@ -15,26 +14,25 @@ export type Props = StateProps & {
   createBucket: typeof BucketActions.createBucket;
 };
 
-export default class BucketCreate extends PureComponent<Props> {
-  render() {
-    const { session, bucket, createBucket } = this.props;
-    const { busy } = session;
-    if (busy) {
-      return <Spinner />;
-    }
-    return (
-      <div>
-        <h1>Create a new bucket</h1>
-        <div className="card">
-          <div className="card-body">
-            <BucketForm
-              session={session}
-              bucket={bucket}
-              onSubmit={({ id, ...attributes }) => createBucket(id, attributes)}
-            />
-          </div>
+export default function BucketCreate({ session, bucket, createBucket }) {
+  const { busy } = session;
+
+  if (busy) {
+    return <Spinner />;
+  }
+
+  return (
+    <div>
+      <h1>Create a new bucket</h1>
+      <div className="card">
+        <div className="card-body">
+          <BucketForm
+            session={session}
+            bucket={bucket}
+            onSubmit={({ id, ...attributes }) => createBucket(id, attributes)}
+          />
         </div>
       </div>
-    );
-  }
+    </div>
+  );
 }
