+++ src/components/bucket/BucketHistory.tsx
@@ -6,7 +6,7 @@ import type {
 } from "../../types";
 import type { Location } from "history";
 
-import React, { PureComponent } from "react";
+import React, { useEffect } from "react";
 
 import * as BucketActions from "../../actions/bucket";
 import * as NotificationActions from "../../actions/notifications";
@@ -14,18 +14,18 @@ import { parseHistoryFilters } from "../../utils";
 import BucketTabs from "./BucketTabs";
 import HistoryTable from "../HistoryTable";
 
-export type OwnProps = {
+type OwnProps = {
   match: BucketRouteMatch;
   location: Location;
 };
 
-export type StateProps = {
+type StateProps = {
   bucket: BucketState;
   capabilities: Capabilities;
   session: SessionState;
 };
 
-export type Props = OwnProps &
+type Props = OwnProps &
   StateProps & {
     listBucketHistory: typeof BucketActions.listBucketHistory;
     listBucketNextHistory: typeof BucketActions.listBucketNextHistory;
@@ -39,54 +39,48 @@ export const onBucketHistoryEnter = (props: Props) => {
   } = match;
   const filters = parseHistoryFilters(location.search);
   if (!session.authenticated) {
-    // We're not authenticated, skip requesting the list of records. This likely
-    // occurs when users refresh the page and lose their session.
     return;
   }
   listBucketHistory(bid, filters);
 };
 
-export default class BucketHistory extends PureComponent<Props> {
-  componentDidMount = () => onBucketHistoryEnter(this.props);
-  componentDidUpdate = (prevProps: Props) => {
-    if (prevProps.location !== this.props.location) {
-      onBucketHistoryEnter(this.props);
-    }
-  };
+export default function BucketHistory(props: Props) {
+  const {
+    match,
+    bucket,
+    capabilities,
+    location,
+    listBucketNextHistory,
+    notifyError,
+  } = props;
 
-  render() {
-    const {
-      match,
-      bucket,
-      capabilities,
-      location,
-      listBucketNextHistory,
-      notifyError,
-    } = this.props;
-    const {
-      params: { bid },
-    } = match;
-    const {
-      history: { entries, loaded, hasNextPage: hasNextHistoryPage },
-    } = bucket;
+  useEffect(() => {
+    onBucketHistoryEnter(props);
+  }, [props.location]);
 
-    return (
-      <div>
-        <h1>
-          History for <b>{bid}</b>
-        </h1>
-        <BucketTabs bid={bid} selected="history" capabilities={capabilities}>
-          <HistoryTable
-            bid={bid}
-            historyLoaded={loaded}
-            history={entries}
-            hasNextHistory={hasNextHistoryPage}
-            listNextHistory={listBucketNextHistory}
-            location={location}
-            notifyError={notifyError}
-          />
-        </BucketTabs>
-      </div>
-    );
-  }
+  const {
+    params: { bid },
+  } = match;
+  const {
+    history: { entries, loaded, hasNextPage: hasNextHistoryPage },
+  } = bucket;
+
+  return (
+    <div>
+      <h1>
+        History for <b>{bid}</b>
+      </h1>
+      <BucketTabs bid={bid} selected="history" capabilities={capabilities}>
+        <HistoryTable
+          bid={bid}
+          historyLoaded={loaded}
+          history={entries}
+          hasNextHistory={hasNextHistoryPage}
+          listNextHistory={listBucketNextHistory}
+          location={location}
+          notifyError={notifyError}
+        />
+      </BucketTabs>
+    </div>
+  );
 }
